Improved capture parsing best preview frames when successful edge
When we detect a successful edge, we keep the associated frame and points. When there are several successful edges (10), we parse frames and keep the best (via checking blur level).

With this, we avoid bad detections when picture was taken. When a picture is taken, there is a latency which causes differences between points found from final picture and points found during preview (because user may have moved during capture or light has changed).

In conclusion, capture is much better and faster than before but we have lost in term of image quality. This is because preview frames pixels density are 2x lower than a taken picture. We must find a better way to improve image quality.